Add support for many custom date formats
We do a lot of normalizing of forms to ensure they are standard, then
use those to build up a date to parse back in at the end. We use this to
get information about the date.

Maybe we could manually specify aspects of the date seeing as we have
most of the information just from the match?